Send adblock status w/ad request
